created new user routes and update watchLists

+++ src/service/watchLists/deleteWatchList.service.ts
@@ -9,13 +9,25 @@ export const deleteWatchListService = async (
   try {
     const user = await getAuth().verifyIdToken(authToken);
 
-    const db = client.db('WatchLists').collection(user.uid);
+    const db = client.db('Reviews').collection('WatchLists');
 
     const id = new ObjectId(uuid);
 
-    const result = await db.deleteOne({ _id: id });
+    const watchList = await db.findOne({ _id: id })
+    
+    if (watchList) {
+      if (watchList.userId === user.uid) {
+        const result = await db.deleteOne({ _id: id });
 
     return result;
+      } else {
+        throw new Error('Not authorized to delete this Watch List')
+      }
+    }
+
+    throw new Error('Watch List not found')
+
+    
   } catch (err) {
     throw err;
   }
